### 2020-06-18 #2360 @whedon
[PRE REVIEW]: htmldate: A Python package to extract publication dates from web pages
**Submitting author:** @adbar (<a href="http://orcid.org/0000-0002-8079-8694">Adrien Barbaresi</a>)
**Repository:** <a href="https://github.com/adbar/htmldate" target ="_blank">https://github.com/adbar/htmldate</a>
**Version:** v0.6.3
