### 2020-06-19 #2360 @danielskatz
@adbar - In the introduction to your paper, please add some text to describe the kinds of research that will use this software.  For example, on your readme, you mention "methodological approach to derive information from web documents in order to build text databases for research (chiefly linguistics and natural language processing). There are web pages for which neither the URL nor the server response provide a reliable way to find out when a document was published or modified."  It would be useful to put some of this concept/info into the paper
