### 2020-06-19 #2360 @danielskatz
Also, note that some of your references are missing DOIs.  After making changes, write a new comment in this thread with ```@whedon generate pdf``` to regenerate the PDF, and use ```@whedon check references``` to check the references.  Also note that whedon sometimes gets references wrong, and makes incorrect suggestions, so please check these.
