### 2020-06-19 #2360 @danielskatz
Finally, please suggest potential reviewers by mentioning them here in this thread (by github username, if they have one, without tagging them with an @). [This list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
