### 2020-06-23 #2360 @danielskatz
👋 @adbar - Can you confirm that you've seen the requests to you?  And when do you think you might have an updated paper ready, along with reviewer suggestions?
