### 2020-06-23 313edaf
more robust encoding handling prior to text search
