### 2020-06-26 #2360 @adbar
Thank you for your feedback! I've seen the requests and I will try to improve the text accordingly. I'm planning to have it done by the end of next week.
