### 2020-07-02 d06c873
Modifying GERMAN_TEXTSEARCH to cover Turkish
The only difference is ".".

German: 1. Januar 1970
Turkish: 1 Ocak 1970
