+++ docs/joss-article/paper.bib
@@ -16,6 +16,7 @@
   pages={327--344},
   year={2017},
   publisher={De Gruyter},
+  doi={10.1515/zgl-2017-0017},
 }
 
 @phdthesis{Barbaresi:2015,
@@ -33,6 +34,7 @@
   year=2016,
   publisher = {Association for Computational Linguistics},
   editor={Cook, Paul and Evert, Stefan and Schäfer, Roland and Stemle, Egon},
+  doi={10.18653/v1/w16-2602},
 }
 
 @misc{Barbaresi:2019t,
@@ -60,7 +62,8 @@
   number={1},
   pages={147--151},
   year={2007},
-  publisher={MIT Press}
+  publisher={MIT Press},
+  doi={10.1162/coli.2007.33.1.147},
 }
 
 @article{Baroni:2009,
@@ -71,7 +74,8 @@
   number={3},
   pages={209--226},
   year={2009},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10579-009-9081-4},
 }
 
 @article{Olston:2010,
