### 2020-07-03 #2360 @adbar
I reviewed the introduction and added the missing DOIs.
