### 2020-07-03 9b6e1e9
Fixed dot character for GERMAN_TEXTSEARCH
