### 2020-07-03 d995d29
Yet another fix for GERMAN_TEXTSEARCH
