### 2020-07-04 #2439 @whedon
[REVIEW]: htmldate: A Python package to extract publication dates from web pages
**Submitting author:** @adbar (<a href="http://orcid.org/0000-0002-8079-8694">Adrien Barbaresi</a>)
**Repository:** <a href="https://github.com/adbar/htmldate" target ="_blank">https://github.com/adbar/htmldate</a>
**Version:** v0.7.0
