### 2020-07-05 #2439 @geoffbacon
Thanks @danielskatz - I will have this done by Friday 10th July (PDT).
