### 2020-07-08 #2439 @proycon
I completed my review, ticking the boxes in the original post (@danielskatz I assume there's some protection layer there confirming that I was the one ticking them right?).

**Overal summary:** I think it's a well written paper and useful software package, suitable for publication in JOSS.

**Some extra comments:**

> Contribution and authorship: Has the submitting author (@adbar) made major contributions to the software? Does the full list           of paper authors seem appropriate and complete?

The 2nd contributor (@DerKozmonaut) I assume to be the person Yannick Kozmus that is explicitly acknowledged in the acknowledgements section, this seems appropriate and complete to me.

I evaluated some of the software's functionality by running it on a few arbitrary URLs:

   * A recent blogpost of mine - https://proycon.anaproy.nl/post/break_free_do_not_be_a_prisoner_of_your_software/ - correctly dated
   * An article on a popular dutch news site - https://www.nu.nl/coronavirus/6063124/nederland-stelt-opnieuw-inreisverbod-in-voor-servie-en-montenegro.html - correctly dated
   * A reddit post - https://www.reddit.com/r/postmarketOS/comments/hc38e3/does_postmarketos_support_wifi_calling/ - not dated at all
   * A forum post (phpBB3) - https://forum.unilang.org/viewtopic.php?f=1&t=58214 - correctly dated
   * A wikipedia article - https://en.wikipedia.org/wiki/Constitution_of_Russia - incorrectly dated? Returns a date but it is not the original publication date (according to wikipedia history) nor the latest edit date. (I also tried with the explicit original_date option)

For the developers, I'd say there's possible room for future improvement in adding some specific heuristics for some big platforms like reddit and wikipedia. (should be easy pickings)
