### 2020-07-08 #2439 @danielskatz
@proycon - thanks!

> I completed my review, ticking the boxes in the original post (@danielskatz I assume there's some protection layer there confirming that I was the one ticking them right?).

no, but the fact that you have written here about the paper being suitable implies that *you* have also checked off all the items, or at least, that you agree with their status as being checked off.
