### 2020-07-08 ba0f35a
Fixed some typos
