### 2020-07-09 ac3ad6d
Declare idiosyncrasies
