### 2020-07-11 1c4ab9b
Declare idiosyncrasies_search
