### 2020-07-11 96094be
Add idiosyncrasies_search
