+++ htmldate/core.py
@@ -689,6 +689,11 @@ def find_date(htmlobject, extensive_search=True, original_date=False, outputform
     text_result = german_text_search(htmlstring, outputformat, max_date)
     if text_result is not None:
         return text_result
+    
+    #idisyncrasies
+    text_result = idiosyncrasies_search(htmlstring, outputformat, max_date)
+    if text_result is not None:
+        return text_result
 
     # last try: URL 2
     if url is not None:
