### 2020-07-11 da2f7f0
test_idiosyncrasies
