### 2020-07-11 03d1435
Fix and improve idiosyncrasies
