+++ htmldate/extractors.py
@@ -136,9 +136,9 @@ TEXT_MONTHS = {'Januar': '01', 'Jänner': '01', 'January': '01', 'Jan': '01', 'O
 TEXT_DATE_PATTERN = re.compile(r'[.:,_/ -]|^[0-9]+$')
 NO_TEXT_DATE_PATTERN = re.compile(r'[0-9]{2}:[0-9]{2}(:| )|\D*[0-9]{4}\D*$')
 
-IDIOSYNCRASIES_EN = regex.compile(r'((U|u)pdated|UPDATED|(P|p)ublished|PUBLISHED)( in| IN)?:? ?[0-9]{1,2}(\.|\/)[0-9]{1,2}(\.|\/)[0-9]{2,4}')
-IDIOSYNCRASIES_TR = regex.compile(r'''(((G|g)üncellen?me|GÜNCELLEN?ME|(Y|y)ayı(n|m)lan?ma|YAYI(N|M)LAN?MA)( tarihi|TARİHİ)?:? ?[0-9]{1,2}(\.|\/)[0-9]{1,2}(\.|\/)[0-9]{2,4})|(
-                                 [0-9]{1,2}(\.|\/)[0-9]{1,2}(\.|\/)[0-9]{2,4}('[a-zA-Z]{2}| tarihinde| TARİHİNDE) (yayı(n|m)landı|YAYI(N|M)LANDI|güncellendi|GÜNCELLENDİ))''')
+IDIOSYNCRASIES_EN = regex.compile(r'[updatedUPDATEDpblishPBLISH]{7,9} ?[inIN]{0,2}:? ?([0-9]{1,2})[./]([0-9]{1,2})[./]([0-9]{2,4})')
+IDIOSYNCRASIES_TR_PREFIXED = regex.compile(r'[güncelenmeGÜNCELENMEyayıaYAYIA]{9,11} [tarihTARİH]{6}?:? ?([0-9]{1,2})[./]([0-9]{1,2})[./]([0-9]{2,4})')
+IDIOSYNCRASIES_TR_SUFFIXED = regex.compile(r'''([0-9]{1,2})[./]([0-9]{1,2})[./]([0-9]{2,4}) ['’tarihndeTARİHNDE]{3,9} [yaınmldYAINMLDgüceiGÜCEİ]{10,11}''')
 
 
 def discard_unwanted(tree):
@@ -398,9 +398,9 @@ def german_text_search(htmlstring, outputformat, max_date):
                 return convert_date(candidate, '%Y-%m-%d', outputformat)
     return None
 
-def idiosyncrasies_search(htmlstring, outputformat, max_date):
-    '''Look for precise German patterns throughout the web page'''
-    match = IDIOSYNCRASIES_EN.search(htmlstring)
+def extract_idiosyncrasy(idiosyncrasy, htmlstring, outputformat, max_date):
+    '''Extract dates in given expression'''
+    match = idiosyncrasy.search(htmlstring)
     if match and len(match.group(3)) in (2, 4):
         try:
             if len(match.group(3)) == 2:
@@ -417,21 +417,18 @@ def idiosyncrasies_search(htmlstring, outputformat, max_date):
             if date_validator(candidate, '%Y-%m-%d', latest=max_date) is True:
                 LOGGER.debug('precise pattern found: %s', match.group(0))
                 return convert_date(candidate, '%Y-%m-%d', outputformat)
-    match = IDIOSYNCRASIES_TR.search(htmlstring)
-    if match and len(match.group(3)) in (2, 4):
-       try:
-           if len(match.group(3)) == 2:
-               candidate = datetime.date(int('20' + match.group(3)),
-                                         int(match.group(2)),
-                                         int(match.group(1)))
-           else:
-               candidate = datetime.date(int(match.group(3)),
[…]
