### 2020-07-13 c724b50
reviewed readme
clarification in credits: libraries → their
