### 2020-07-13 8a63d81
reviewed specific extraction patterns
