### 2020-07-13 2f19338
Update regex for idiosyncrasies
For Turkish, the regex matches similar to the following;

güncelleme: <date>
güncellenme: <date>
yayımlama: <date>
yayımlanma: <date>
yayınlama: <date>
yayımlama: <date>
güncelleme tarihi: <date>
güncellenme tarihi: <date>
yayımlama tarihi: <date>
yayımlanma tarihi: <date>
yayınlama tarihi: <date>
yayımlama tarihi: <date>
<date>'de yayımlandı [the suffix for date differs grammatically, see <https://quarry.wmflabs.org/query/46220> for complete example.]
<date>'de yayınlandı
<date>'de güncellendi
<date> tarihinde yayımlandı
<date> tarihinde yayınlandı
<date> tarih
[…]
