### 2020-07-13 55a028c
Remove indentation in regex
