### 2020-07-13 #2439 @geoffbacon
Finished my review. Like @proycon,  I find htmldate to be a useful utility library with a clear motivation. I particularly appreciated the evaluation suite that tests the library against numerous other existing solutions.
