### 2020-07-13 #2439 @danielskatz
@adbar- to finish this process, I've suggested some changes to the paper in https://github.com/adbar/htmldate/pull/15
Please also fix the references, for example the booktitle in Hamborg et al. has some incorrect cases, as does at least on python (that should be Python).  Please check all the references carefully, and use {}s in the bibtex to protect cases.

Then use `@whedon generate pdf` to regenerate the PDF to check.

Once you are happy with this, please
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
