### 2020-07-14 0038735
Update extract_idiosyncrasy
