+++ htmldate/extractors.py
@@ -403,7 +403,10 @@ def extract_idiosyncrasy(idiosyncrasy, htmlstring, outputformat, max_date):
     '''Extract dates in given expression'''
     match = idiosyncrasy.search(htmlstring)
     groups = [0, 1, 2, 3] if match and match.group(3) else [] #because len(None) has no len
-    groups = [0, 4, 5, 6] if match and match.group(6) else groups #because len(None) has no len
+    try:
+        groups = [0, 4, 5, 6] if match and match.group(6) else groups #because len(None) has no len
+    except IndexError:
+        pass
     if match and groups: #because len(None) has no len
         if len(match.group(groups[3])) in (2, 4):
             try:
