### 2020-07-14 75d792c
Remove newline from regex
