### 2020-07-14 #2439 @adbar
@proycon @geoffbacon thank you for your feedback!

@danielskatz thank you for the suggestions and the guidelines, I'll proceed as instructed above.
