### 2020-07-15 2aafa1e
multilingual extraction: code linted and simplified
