### 2020-07-15 cf9f708
tests fix: make examples part of pytest routine
