+++ tests/unit_tests.py
@@ -28,7 +28,7 @@ except ImportError:
 
 from htmldate.cli import examine, parse_args
 from htmldate.core import compare_reference, find_date, search_page, search_pattern, select_candidate, try_ymd_date
-from htmldate.extractors import custom_parse, external_date_parser, extract_partial_url_date, regex_parse_de, regex_parse_en
+from htmldate.extractors import custom_parse, external_date_parser, extract_partial_url_date, regex_parse
 from htmldate.settings import LATEST_POSSIBLE
 from htmldate.utils import fetch_url, load_html
 from htmldate.validators import convert_date, date_validator, output_format_validator
@@ -401,17 +401,17 @@ def test_candidate_selection(original_date=False, max_date=LATEST_POSSIBLE):
 
 def test_regex_parse():
     '''test date extraction using rules and regular expressions'''
-    assert regex_parse_de('3. Dezember 2008') is not None
-    assert regex_parse_de('33. Dezember 2008') is None
-    assert regex_parse_de('3 Aralık 2008 Çarşamba') is not None
-    assert regex_parse_de('3 Aralık 2008') is not None
-    assert regex_parse_en('Tuesday, March 26th, 2019') is not None
-    assert regex_parse_en('March 26, 2019') is not None
-    assert regex_parse_en('3rd Tuesday in March') is None
-    assert regex_parse_en('Mart 26, 2019') is not None
-    assert regex_parse_en('Salı, Mart 26, 2019') is not None
-    assert regex_parse_en('3/14/2016') is not None
-    assert regex_parse_en('36/14/2016') is None
+    assert regex_parse('3. Dezember 2008') is not None
+    assert regex_parse('33. Dezember 2008') is None
+    assert regex_parse('3 Aralık 2008 Çarşamba') is not None
+    assert regex_parse('3 Aralık 2008') is not None
+    assert regex_parse('Tuesday, March 26th, 2019') is not None
+    assert regex_parse('March 26, 2019') is not None
+    assert regex_parse('3rd Tuesday in March') is None
+    assert regex_parse('Mart 26, 2019') is not None
+    assert regex_parse('Salı, Mart 26, 2019') is not None
+    assert regex_parse('3/14/2016') is not None
+    assert regex_parse('36/14/2016') is None
     assert custom_parse('12122004', OUTPUTFORMAT, False, LATEST_POSSIBLE) is None
     assert custom_parse('20041212', OUTPUTFORMAT, False, LATEST_POSSIBLE) is not None
     assert custom_parse('20041212', OUTPUTFORMAT, True, LATEST_POSSIBLE) is not None
@@ -543,7 +543,7 @@ def test_download():
     assert examine(teststring, False) is None
 
 
-def readme_examples():
+def test_readme_examples():
     '''Test README example for consistency'''
     assert find_date(load_mock_page('http://blog.python.org/2016/12/python-360-is-now-available.html')) == '2016-12-23'
     assert find_date(load_mock_page('https://creativecommons.org/about/'), extensive_search=False) is None
@@ -552,7 +552,7 @@ def readme_examples():
     mytree = html.fromstring('<html><body><span class="entry-date">July 12th, 2016</span></body></html>')
     assert find_date(mytree) == '2016-07-12'
     assert find_date(load_mock_page('https://www.gnu.org/licenses/gpl-3.0.en.html'), outputformat='%d %B %Y') == '18 November 2016'
-    assert find_date(load_mock_page('https://netzpolitik.org/2016/die-cider-connection-abmahnungen-gegen-nutzer-von-creative-commons-bildern/')) == '2019-06-24'
[…]
