+++ README.rst
@@ -41,7 +41,7 @@ htmldate: find the publication date of web pages
 
 |
 
-Find original and updated publication dates of any web page. All the steps needed from web page download to HTML parsing, scraping and text analysis are included.
+Find original and updated publication dates of any web page. From the command-line or within Python, all the steps needed from web page download to HTML parsing, scraping, and text analysis are included.
 
 In a nutshell, with Python:
 
@@ -64,7 +64,7 @@ On the command-line:
 Features
 --------
 
-*htmldate* finds original and updated publication dates of web pages using heuristics on HTML code and linguistic patterns. URLs, HTML files or HTML trees are given as input, the library outputs a date string in the desired format. It provides following ways to date a HTML document:
+*htmldate* finds original and updated publication dates of web pages using heuristics on HTML code and linguistic patterns. URLs, HTML files, or HTML trees are given as input. It provides following ways to date a HTML document:
 
 1. **Markup in header**: common patterns are used to identify relevant elements (e.g. ``link`` and ``meta`` elements) including `Open Graph protocol <http://ogp.me/>`_ attributes and a large number of CMS idiosyncrasies
 2. **HTML code**: The whole document is then searched for structural markers: ``abbr`` and ``time`` elements as well as a series of attributes (e.g. ``postmetadata``)
@@ -73,13 +73,14 @@ Features
   - in ``fast`` mode the HTML page is cleaned and precise patterns are targeted
   - in ``extensive`` mode all potential dates are collected and a disambiguation algorithm determines the best one
 
-The module returns a date if a valid cue could be found in the document, corresponding to either the last update (default) or the original publishing statement. The output string defaults to `ISO 8601 YMD format <https://en.wikipedia.org/wiki/ISO_8601>`_.
+The output is thouroughly verified in terms of plausibility and adequateness and the library outputs a date string, corresponding to either the last update or the original publishing statement (the default), in the desired format (defaults to `ISO 8601 YMD format <https://en.wikipedia.org/wiki/ISO_8601>`_).
 
--  Should be compatible with all common versions of Python 3
--  Output thoroughly verified in terms of plausibility and adequateness
+-  Should be compatible with all recent versions of Python (currently 3.4 to 3.9)
 -  Designed to be computationally efficient and used in production on millions of documents
 -  Batch processing of a list of URLs
 -  Switch between original and updated date
+
+Markup-based extraction is multilingual by nature, text-based refinements for better coverage currently support German, English and Turkish.
 
 
 Performance
@@ -169,8 +170,6 @@ Although the time delta between original publication and "last modified" info is
 
 .. code-block:: python
 
-    >>> find_date('https://netzpolitik.org/2016/die-cider-connection-abmahnungen-gegen-nutzer-von-creative-commons-bildern/') # default setting
-    '2019-06-24'
     >>> find_date('https://netzpolitik.org/2016/die-cider-connection-abmahnungen-gegen-nutzer-von-creative-commons-bildern/', original_date=True) # modified behavior
     '2016-06-23'
 
@@ -223,7 +222,7 @@ If the date is nowhere to be found, it might be worth considering `carbon dating
 Author
 ------
 
-This effort is part of a methodological approach to derive information from web documents in order to build text databases for research (chiefly linguistics and natural language processing). There are web pages for which neither the URL nor the server response provides a reliable way to find out when a document was published or modified. For more information:
+This effort is part of a methodological approach to derive information from web documents in order to build text databases for research (chiefly linguistics and natural language processing). There are web pages for which neither the URL nor the server response provide a reliable way to find out when a document was published or modified. For more information:
 
 .. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.3459599.svg
    :target: https://doi.org/10.5281/zenodo.3459599
