### 2020-07-16 2d05cc0
Remove duplicate month names and some code readibility
