### 2020-07-27 a70b924
discard invalid numeral expressions: DDDD-DDDD
