### 2020-07-27 7c25c13
implemented upper limit for number of potential candidates
