### 2020-07-27 b7666df
perf: move compiled regexes out of loop
