+++ htmldate/core.py
@@ -16,10 +16,19 @@ from functools import lru_cache, partial
 from lxml import etree, html
 
 # own
-from .extractors import (ADDITIONAL_EXPRESSIONS, DATE_EXPRESSIONS,
-                         discard_unwanted, extract_url_date,
+from .extractors import (discard_unwanted, extract_url_date,
                          extract_partial_url_date, idiosyncrasies_search,
-                         json_search, timestamp_search, try_ymd_date)
+                         json_search, timestamp_search, try_ymd_date,
+                         ADDITIONAL_EXPRESSIONS, DATE_EXPRESSIONS,
+                         COPYRIGHT_PATTERN, COPYRIGHT_YEAR, COPYRIGHT_CATCH,
+                         THREE_PATTERN, THREE_YEAR, THREE_CATCH,
+                         THREE_LOOSE_PATTERN, THREE_LOOSE_YEAR, THREE_LOOSE_CATCH,
+                         SELECT_YMD_PATTERN, SELECT_YMD_YEAR, YMD_CATCH, YMD_YEAR,
+                         DATESTRINGS_PATTERN, DATESTRINGS_YEAR, DATESTRINGS_CATCH,
+                         SLASHES_PATTERN, SLASHES_YEAR,
+                         YYYYMM_PATTERN, YYYYMM_YEAR, YYYYMM_CATCH,
+                         MMYYYY_PATTERN, MMYYYY_YEAR,
+                         SIMPLE_PATTERN, SIMPLE_YEAR, SIMPLE_CATCH)
 from .settings import HTML_CLEANER, MAX_POSSIBLE_CANDIDATES
 from .utils import load_html
 from .validators import (check_extracted_reference, compare_values,
@@ -400,42 +409,31 @@ def search_page(htmlstring, outputformat, original_date, max_date):
     # copyright symbol
     LOGGER.debug('looking for copyright/footer information')
     copyear = 0
-    pattern = re.compile(r'(?:©|\&copy;|Copyright|\(c\))\D*([12][0-9]{3})\D')
-    yearpat = re.compile(r'^\D?([12][0-9]{3})')
-    catch = re.compile(r'^\D?([12][0-9]{3})')
-    bestmatch = search_pattern(htmlstring, pattern, catch, yearpat, original_date, max_date)
+    bestmatch = search_pattern(htmlstring, COPYRIGHT_PATTERN, COPYRIGHT_CATCH, COPYRIGHT_YEAR, original_date, max_date)
     if bestmatch is not None:
         LOGGER.debug('Copyright detected: %s', bestmatch.group(0))
         pagedate = '-'.join([bestmatch.group(0), '07', '01'])
         if date_validator(bestmatch.group(0), '%Y', latest=max_date) is True:
-            LOGGER.debug('date found for copyright/footer pattern "%s": %s', pattern, pagedate)
+            LOGGER.debug('date found for copyright/footer pattern "%s": %s', COPYRIGHT_PATTERN, pagedate)
             copyear = int(bestmatch.group(0))
             # return convert_date(pagedate, '%Y-%m-%d', outputformat)
 
     # 3 components
     LOGGER.debug('3 components')
     # target URL characteristics
-    pattern = re.compile(r'/([0-9]{4}/[0-9]{2}/[0-9]{2})[01/]')
-    yearpat = re.compile(r'^\D?([12][0-9]{3})')
-    catch = re.compile(r'([0-9]{4})/([0-9]{2})/([0-9]{2})')
-    bestmatch = search_pattern(htmlstring, pattern, catch, yearpat, original_date, max_date)
-    result = filter_ymd_candidate(bestmatch, pattern, copyear, outputformat, max_date)
[…]
