### 2020-07-28 9f222b7
fix: stricter extensive patterns
