+++ htmldate/extractors.py
@@ -167,14 +167,14 @@ DATESTRINGS_YEAR = re.compile(r'^\D?([12][0-9]{3})')
 DATESTRINGS_CATCH = re.compile(r'([12][0-9]{3})([01][0-9])([0-3][0-9])')
 SLASHES_PATTERN = re.compile(r'\D([0-3]?[0-9][/.][01]?[0-9][/.][019][0-9])\D')
 SLASHES_YEAR = re.compile(r'([0-9]{2})$')
-YYYYMM_PATTERN = re.compile(r'\D([0-9]{4}[/.-][0-9]{2})\D')
+YYYYMM_PATTERN = re.compile(r'\D([0-9]{4}[/.-][01][0-9])\D')
 YYYYMM_YEAR = re.compile(r'^\D?([12][0-9]{3})')
-YYYYMM_CATCH = re.compile(r'([0-9]{4})[/.-]([0-9]{2})')
+YYYYMM_CATCH = re.compile(r'([0-9]{4})[/.-]([01][0-9])')
 MMYYYY_PATTERN = re.compile(r'\D([0-3]?[0-9][/.-][0-9]{4})\D')
 MMYYYY_YEAR = re.compile(r'([12][0-9]{3})\D?$')
-SIMPLE_PATTERN = re.compile(r'\D([12][0-9]{3})\D')
-SIMPLE_YEAR = re.compile(r'^\D?([12][0-9]{3})')
-SIMPLE_CATCH = re.compile(r'^\D?([12][0-9]{3})')
+SIMPLE_PATTERN = re.compile(r'\D(199[0-9]|2[0-2][0-9]{2})\D')
+SIMPLE_YEAR = re.compile(r'^\D?(199[0-9]|2[0-2][0-9]{2})')
+SIMPLE_CATCH = re.compile(r'^\D?(199[0-9]|2[0-2][0-9]{2})')
 
 
 def discard_unwanted(tree):
