### 2020-07-28 9930449
fix: shorter xpaths and higher number of candidates
