+++ htmldate/settings.py
@@ -23,7 +23,7 @@ LATEST_POSSIBLE = datetime.date.today()
 MAX_YEAR = datetime.date.today().year
 
 # set an upper limit to the number of candidates
-MAX_POSSIBLE_CANDIDATES = 50
+MAX_POSSIBLE_CANDIDATES = 100
 
 # HTML_CLEANER config # http://lxml.de/api/lxml.html.clean.Cleaner-class.html
 HTML_CLEANER = Cleaner()
