### 2020-07-28 6b5aae6
added published/modified switch for dates in JSON
