### 2020-07-29 e53e324
added minimum date parameter
