+++ htmldate/core.py
@@ -33,7 +33,7 @@ from .settings import HTML_CLEANER, MAX_POSSIBLE_CANDIDATES
 from .utils import load_html
 from .validators import (check_extracted_reference, compare_values,
                          convert_date, date_validator, filter_ymd_candidate,
-                         get_max_date, output_format_validator,
+                         get_min_date, get_max_date, output_format_validator,
                          plausible_year_filter)
 
 # TODO:
@@ -45,7 +45,7 @@ LOGGER = logging.getLogger(__name__)
 
 
 @lru_cache(maxsize=32)
-def examine_date_elements(tree, expression, outputformat, extensive_search, max_date):
+def examine_date_elements(tree, expression, outputformat, extensive_search, min_date, max_date):
     """Check HTML elements one by one for date expressions"""
     try:
         elements = tree.xpath(expression)
@@ -71,14 +71,14 @@ def examine_date_elements(tree, expression, outputformat, extensive_search, max_
             elem, pretty_print=False, encoding='unicode').translate(
                 {ord(c): None for c in '\n\t\r'}
             ).strip()[:100])
-        attempt = try_ymd_date(toexamine, outputformat, extensive_search, max_date)
+        attempt = try_ymd_date(toexamine, outputformat, extensive_search, min_date, max_date)
         if attempt is not None:
             return attempt
     # catchall
     return None
 
 
-def examine_header(tree, outputformat, extensive_search, original_date, max_date):
+def examine_header(tree, outputformat, extensive_search, original_date, min_date, max_date):
     """
     Parse header elements to find date cues
 
@@ -96,11 +96,17 @@ def examine_header(tree, outputformat, extensive_search, original_date, max_date
         Look for original date (e.g. publication date) instead of most recent
         one (e.g. last modified, updated time)
     :type original_date: boolean
+    :param min_date:
+        Set the earliest acceptable date manually (YYYY-MM-DD format)
+    :type min_date: string
+    :param max_date:
+        Set the latest acceptable date manually (YYYY-MM-DD format)
+    :type max_date: string
     :return: Returns a valid date expression as a string, or None
 
     """
     headerdate, reserve = None, None
-    tryfunc = partial(try_ymd_date, outputformat=outputformat, extensive_search=extensive_search, max_date=max_date)
+    tryfunc = partial(try_ymd_date, outputformat=outputformat, extensive_search=extensive_search, min_date=min_date, max_date=max_date)
     # loop through all meta elements
     for elem in tree.xpath('//meta'):
         # safeguard
@@ -166,14 +172,14 @@ def examine_header(tree, outputformat, extensive_search, original_date, max_date
                 ) and headerdate is None:
                 LOGGER.debug('examining meta itemprop: %s', html.tostring(elem, pretty_print=False, encoding='unicode').strip())
                 if 'datetime' in elem.attrib:
-                    headerdate = try_ymd_date(elem.get('datetime'), outputformat, extensive_search, max_date)
+                    headerdate = try_ymd_date(elem.get('datetime'), outputformat, extensive_search, min_date, max_date)
                 elif 'content' in elem.attrib:
                     headerdate = tryfunc(elem.get('content'))
             # override
             elif elem.get('itemprop').lower() == 'datemodified' and original_date is False:
                 LOGGER.debug('examining meta itemprop: %s', html.tostring(elem, pretty_print=False, encoding='unicode').strip())
                 if 'datetime' in elem.attrib:
-                    attempt = try_ymd_date(elem.get('datetime'), outputformat, extensive_search, max_date)
[…]

[…]
