### 2020-07-29 #2439 @danielskatz
👋 @adbar - I haven't seen an update in 15 days, and there's certainly nothing major left to do.  How are you proceeding on these final steps?
