### 2020-07-29 ae1f6dd
add minimum date as CLI option
