+++ docs/joss-article/paper.bib
@@ -27,7 +27,7 @@
 }
 
 @inproceedings{Barbaresi:2016,
-  title={Efficient construction of metadata-enhanced web corpora},
+  title={{Efficient construction of metadata-enhanced web corpora}},
   author={Barbaresi, Adrien},
   booktitle={{Proceedings of the 10th Web as Corpus Workshop}},
   pages={7--16},
@@ -46,9 +46,9 @@
 
 @InProceedings{HamborgEtAl:2017,
   author     = {Hamborg, Felix and Meuschke, Norman and Breitinger, Corinna and Gipp, Bela},
-  title      = {news-please: A Generic News Crawler and Extractor},
+  title      = {{news-please: A Generic News Crawler and Extractor}},
   year       = {2017},
-  booktitle  = {Proceedings of the 15th International Symposium of Information Science},
+  booktitle  = {{Proceedings of the 15th International Symposium of Information Science}},
   location   = {Berlin},
   editor     = {Gaede, Maria and Trkulja, Violeta and Petra, Vivien},
   pages      = {218--223},
@@ -100,7 +100,7 @@
 
 @misc{articleDateExtractor,
   author = {Geva, Ran et al.},
-  title = {articleDateExtractor: Automatically extracts and normalizes an online article or blog post publication date},
+  title = {{articleDateExtractor: Automatically extracts and normalizes an online article or blog post publication date},
   year = {2018},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -109,7 +109,7 @@
 
 @misc{dateguesser,
   author = {Carroll, Colin D. and Valiukas, Linas et al.},
-  title = {Date Guesser: A library to extract a publication date from a web page, along with a measure of the accuracy},
+  title = {{Date Guesser: A library to extract a publication date from a web page, along with a measure of the accuracy}},
   year = {2019},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -118,7 +118,7 @@
 
 @misc{goose3,
   author = {Grangier, Xavier and Barrus, Tyler and Sidorov, Ilya et al.},
-  title = {Goose3: A Python 3 compatible version of goose},
+  title = {{Goose3: A Python 3 compatible version of goose}},
   year = {2019},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -127,7 +127,7 @@
 
 @misc{newspaper,
   author = {Ou-Yang, Lucas and Prezument, Yuri et al.},
-  title = {Newspaper: News, full-text, and article metadata extraction in Python 3},
+  title = {{Newspaper: News, full-text, and article metadata extraction in Python 3}},
   year = {2019},
   publisher = {GitHub},
   journal = {GitHub repository},
