+++ docs/joss-article/paper.bib
@@ -100,7 +100,7 @@
 
 @misc{articleDateExtractor,
   author = {Geva, Ran et al.},
-  title = {{articleDateExtractor: Automatically extracts and normalizes an online article or blog post publication date},
+  title = {{articleDateExtractor: Automatically extracts and normalizes an online article or blog post publication date}},
   year = {2018},
   publisher = {GitHub},
   journal = {GitHub repository},
