+++ setup.py
@@ -18,8 +18,8 @@ packages = ['htmldate']
 extras = {
     'all': [
         'ciso8601 >= 2.1.3',
-        'dateparser >= 0.7.4',  # 0.5.0 could be faster
-        'regex >= 2020.5.14',
+        'dateparser >= 0.7.6',  # 0.5.0 could be faster
+        'regex >= 2020.7.14',
         ],
 }
 
@@ -30,7 +30,7 @@ def readme():
 
 setup(
     name='htmldate',
-    version='0.6.3',
+    version='0.7.0',
     description='Fast and robust extraction of original and updated publication dates from web pages.',
     long_description=readme(),
     classifiers=[
@@ -68,10 +68,10 @@ setup(
     python_requires='>=3.4',
     install_requires=[
         'lxml == 4.3.5; python_version == "3.4"',
-        'lxml >= 4.5.1; python_version > "3.4"',
+        'lxml >= 4.5.2; python_version > "3.4"',
         'python-dateutil >= 2.8.1',
         'requests == 2.21.0; python_version == "3.4"',
-        'requests >= 2.23.0; python_version > "3.4"',
+        'requests >= 2.24.0; python_version > "3.4"',
     ],
     extras_require=extras,
     entry_points = {
