+++ htmldate/__init__.py
@@ -7,7 +7,7 @@ __title__ = 'htmldate'
 __author__ = 'Adrien Barbaresi'
 __license__ = 'GNU GPL v3'
 __copyright__ = 'Copyright 2017-2020, Adrien Barbaresi'
-__version__ = '0.6.3'
+__version__ = '0.7.0'
 
 
 import logging
