### 2020-07-29 1547051
fixed docs compilation errors
