### 2020-07-29 #2439 @adbar
@proycon The issue with Wikipedia is now solved. Reddit doesn't currently display dates that are extractable in a straightforward way, I'll keep this case in mind.
