### 2020-07-30 ea22de1
small change in paper wording
