### 2020-07-30 #2439 @danielskatz
@adbar - can you merge https://github.com/adbar/htmldate/pull/20 as well?
